+++ FireAxe/Models/Curves/Arc.cs
@@ -8,25 +8,25 @@ namespace FireAxe.Models.Curves
 {
     public class Arc : Curve
     {
-        public override double RecommendedInterval => (end-start)*64;
+        public override float RecommendedInterval => (end-start)*64;
             
         public override IEnumerable<(Double3m, Double3m)> BoundingBoxes => throw new NotImplementedException();
 
         private Round baseRound;
-        private double start;
-        private double end;
+        private float start;
+        private float end;
         public Arc(Round baseRound)
         {
             this.baseRound = baseRound;
         }
-        public Arc(Double3m origin, Double3m normal, double radius,double start,double end)
+        public Arc(Double3m origin, Double3m normal, float radius,float start,float end)
         {
             baseRound = new Round(origin, normal, radius);
             this.start = start;
             this.end = end;
         }
 
-        public override Double3m GetPoint(double T)
+        public override Double3m GetPoint(float T)
         {
             return(baseRound.GetPoint(T * (end-start) + start));
         }
