+++ FireAxe/Models/Points/Point.cs
@@ -5,9 +5,9 @@
     /// </summary>
     public class Point
     {
-        public double X { get; set; }
-        public double Y { get; set; }
-        public double Z { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
+        public float Z { get; set; }
